requirements/base.txt
django-filter==21.1
django-model-utils==4.2.0
django-picklefield==3.0.1
django-select2==7.9.0
django-tables2==2.4.1
django-widget-tweaks==1.4.12
django==4.0.1
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
coreapi==2.3.3
gunicorn==20.1.0
pillow==9.0.0
python-decouple==3.5
pytz==2021.3
